pyproject.toml
[tool.poetry]
name = "notified"
version = "0.1.1"
description = ""
authors = ["Oleskii Osadchul <aleosd@gmail.com>"]
readme = "README.md"
packages = [{include = "notified"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9"
nanos = "^0.1.4"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.2"
ruff = "^0.8.1"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
